refactor: 일회성 타입을 api 함수 파일 내부에 정의
