+++ src/app/lib/comment/patchComment.ts
@@ -1,5 +1,10 @@
-import { PatchTaskCommentRequest } from '@/app/types/comment';
 import instance from '../instance';
+
+interface PatchTaskCommentRequest {
+  taskId: number;
+  commentId: number;
+  content: string;
+}
 
 // 할 일 댓글 수정
 const editTaskComment = async ({
