+++ src/app/lib/task/deleteTask.ts
@@ -1,5 +1,17 @@
-import { DeleteRecurringRequest, DeleteTaskRequest } from '@/app/types/task';
 import axios from 'axios';
+
+interface DeleteTaskRequest {
+  groupId: number;
+  taskListId: number;
+  taskId: number;
+}
+
+interface DeleteRecurringRequest {
+  groupId: number;
+  taskListId: number;
+  taskId: number;
+  recurringId: number;
+}
 
 // 할 일 삭제
 export const deleteTask = async ({
