+++ src/app/lib/task/postRecurringTask.ts
@@ -1,5 +1,11 @@
-import { PostRecurringTaskRequest } from '@/app/types/recurring';
+import { RecurringTaskDataBody } from '@/app/types/task';
 import axios from 'axios';
+
+export interface PostRecurringTaskRequest {
+  groupId: number;
+  taskListId: number;
+  data: RecurringTaskDataBody;
+}
 
 // 반복 할 일 생성
 export const createRecurringTask = async ({
