+++ src/app/lib/tasklist/getTaskList.ts
@@ -1,5 +1,21 @@
-import { GetTaskListRequest, GetTaskListResponse } from '@/app/types/taskList';
+import { Task } from '@/app/types/task';
 import instance from '../instance';
+
+interface GetTaskListRequest {
+  groupId: number;
+  id: number;
+  date: string;
+}
+
+interface GetTaskListResponse {
+  displayIndex: number;
+  groupId: number;
+  updatedAt: string;
+  createdAt: string;
+  name: string;
+  id: number;
+  tasks: Task[];
+}
 
 // 할 일 목록 조회
 const getTaskList = async ({
