+++ src/app/lib/tasklist/patchTaskList.ts
@@ -1,9 +1,25 @@
-import {
-  PatchTaskListRequest,
-  PatchTaskListResponse,
-} from '@/app/types/taskList';
-import { PatchTaskOrderRequest } from '@/app/types/task';
 import instance from '../instance';
+
+interface PatchTaskListRequest {
+  groupId: number;
+  id: number;
+  name: string;
+}
+
+interface PatchTaskListResponse {
+  displayIndex: number;
+  groupId: number;
+  updatedAt: string;
+  createdAt: string;
+  name: string;
+  id: number;
+}
+
+interface PatchTaskListOrderRequest {
+  groupId: number;
+  id: number;
+  displayIndex: number;
+}
 
 // 할 일 목록 이름 변경
 export const editTaskList = async ({
@@ -19,11 +35,11 @@ export const editTaskList = async ({
 };
 
 // 할 일 목록 순서 변경
-export const editTaskOrder = async ({
+export const editTaskListOrder = async ({
   groupId,
   id,
   displayIndex,
-}: PatchTaskOrderRequest) => {
+}: PatchTaskListOrderRequest) => {
   const res = await instance.patch(`groups/${groupId}/task-lists/${id}/order`, {
     displayIndex,
   });
