+++ src/app/lib/tasklist/postTaskList.ts
@@ -1,8 +1,18 @@
-import {
-  PostTaskListRequest,
-  PostTaskListResponse,
-} from '@/app/types/taskList';
 import instance from '../instance';
+
+interface PostTaskListRequest {
+  groupId: number;
+  name: string;
+}
+
+interface PostTaskListResponse {
+  displayIndex: number;
+  groupId: number;
+  updatedAt: string;
+  createdAt: string;
+  name: string;
+  id: number;
+}
 
 // 할 일 목록 생성
 const createTaskList = async ({
