--- src/app/types/recurring.ts
@@ -1,7 +1,0 @@
-import { RecurringTaskDataBody } from './task';
-
-export interface PostRecurringTaskRequest {
-  groupId: number;
-  taskListId: number;
-  data: RecurringTaskDataBody;
-}
