+++ src/app/types/task.ts
@@ -72,55 +72,4 @@ interface PostTaskRequest {
   data: RecurringTaskData;
 }
 
-interface GetTasksRequest {
-  groupId: number;
-  taskListId: number;
-  date?: string;
-}
-
-interface GetTaskDetailRequest {
-  groupId: number;
-  taskListId: number;
-  taskId: number;
-}
-
-interface PatchTaskRequest {
-  groupId: number;
-  taskListId: number;
-  taskId: number;
-  name?: string;
-  description?: string;
-  done?: boolean;
-}
-
-interface PatchTaskOrderRequest {
-  groupId: number;
-  taskListId: number;
-  id: number;
-  displayIndex: number;
-}
-
-interface DeleteTaskRequest {
-  groupId: number;
-  taskListId: number;
-  taskId: number;
-}
-
-interface DeleteRecurringRequest {
-  groupId: number;
-  taskListId: number;
-  taskId: number;
-  recurringId: number;
-}
-
-export type {
-  Task,
-  RecurringTaskDataBody,
-  PostTaskRequest,
-  GetTasksRequest,
-  GetTaskDetailRequest,
-  PatchTaskRequest,
-  PatchTaskOrderRequest,
-  DeleteTaskRequest,
-  DeleteRecurringRequest,
-};
+export type { Task, RecurringTaskData, RecurringTaskDataBody, PostTaskRequest };
